requirements.txt
Pillow==10.0.1
numpy==1.26.0
pygraphviz==1.11
SPARQLWrapper==2.0.0
boto3==1.28.61
botocore==1.31.61
face-recognition==1.3.0
alive-progress==3.1.5
python-dotenv==1.0.0